# api/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Famiphoto API
  version: 0.0.0
tags:
  - name: admin
  - name: auth
paths:
  /admin/users:
    post:
      operationId: AdminUserManagement_createUser
      summary: ユーザアカウントを作成します
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin.CreateUserResponse'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Admin.CreateUserRequest'
      security:
        - ApiKeyAuth: []
  /admin/users/{userId}:
    delete:
      operationId: AdminUserManagement_deleteUser
      summary: ユーザアカウントを無効化します
      parameters:
        - name: userId
          in: path
          required: true
          description: 削除対象のユーザID
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - admin
      security:
        - ApiKeyAuth: []
  /assets/previews/{photoId}:
    get:
      operationId: Assets_getPreview
      summary: プレビュー画像を取得します
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
  /assets/thumbnails/{photoId}:
    get:
      operationId: Assets_getThumbnail
      summary: サムネイル画像を取得します
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
  /auth/me:
    get:
      operationId: Auth_getMe
      summary: ログイン中ユーザの情報を取得します。
      description: ログイン中かどうかの判定にも使います。
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auth.MeResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - auth
      security:
        - ApiKeyAuth: []
  /auth/sign_in:
    post:
      operationId: Auth_postSignIn
      summary: サインイン
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auth.SignInResponse'
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth.SignInRequest'
      security:
        - {}
  /auth/sign_out:
    post:
      operationId: Auth_postSignOut
      summary: サインアウトします
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - auth
      security:
        - ApiKeyAuth: []
  /health:
    get:
      operationId: Health_getHealth
      summary: ヘルスチェック
      description: APIサーバーの状態を確認する
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
      security:
        - {}
  /photos:
    get:
      operationId: Photos_getPhotoList
      summary: 写真一覧を取得します。
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 30
          explode: false
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: 0
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photos.PhotoListResponse'
      security:
        - ApiKeyAuth: []
  /photos/{photoId}:
    get:
      operationId: Photos_getPhoto
      summary: 写真情報を取得します。
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photos.Photo'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
components:
  schemas:
    Admin.CreateUserRequest:
      type: object
      required:
        - userId
        - password
      properties:
        userId:
          type: string
          description: 取得したいUserIDを指定します
        password:
          type: string
          description: 設定したいパスワード
        isAdmin:
          type: boolean
          description: 管理者権限を付与するか
          default: false
    Admin.CreateUserResponse:
      type: object
      required:
        - userId
        - isAdmin
      properties:
        userId:
          type: string
        isAdmin:
          type: boolean
    Auth.MeResponse:
      type: object
      required:
        - userId
        - isAdmin
      properties:
        userId:
          type: string
        isAdmin:
          type: boolean
    Auth.SignInRequest:
      type: object
      required:
        - userId
        - password
      properties:
        userId:
          type: string
          description: 取得したいUserIDを指定します
        password:
          type: string
          description: 設定したいパスワード
    Auth.SignInResponse:
      type: object
      required:
        - userId
        - isAdmin
      properties:
        userId:
          type: string
        isAdmin:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - statusCode
        - errorCode
      properties:
        statusCode:
          type: integer
          description: HTTPステータスコード
        errorCode:
          type: string
          description: APIが定義するエラーコード
        errorMessage:
          type: string
          description: システムエラーメッセージ詳細(本番環境はなし)
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          default: ok
    Photos.ExifData:
      type: object
      required:
        - make
        - model
        - serialNumber
        - dateTimeOriginal
        - createDate
        - subsecTimeOriginal
        - timezoneOffset
        - exposureTime
        - fNumber
        - iso
        - focalLength
        - focalLengthIn35mm
        - exposureProgram
        - exposureCompensation
        - meteringMode
        - flash
        - lensMake
        - lensModel
        - lensSerialNumber
        - width
        - height
        - colorSpace
        - whiteBalance
        - orientation
        - software
        - firmware
      properties:
        make:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        dateTimeOriginal:
          type: integer
          format: int64
        createDate:
          type: integer
          format: int64
        subsecTimeOriginal:
          type: string
        timezoneOffset:
          type: string
        exposureTime:
          type: number
          format: double
        fNumber:
          type: number
          format: double
        iso:
          type: integer
          format: int64
        focalLength:
          type: number
          format: double
        focalLengthIn35mm:
          type: number
          format: double
        exposureProgram:
          type: integer
          format: int64
        exposureCompensation:
          type: number
          format: double
        meteringMode:
          type: integer
          format: int64
        flash:
          type: integer
          format: int64
        lensMake:
          type: string
        lensModel:
          type: string
        lensSerialNumber:
          type: string
        width:
          type: integer
          format: int64
        height:
          type: integer
          format: int64
        colorSpace:
          type: integer
          format: int64
        whiteBalance:
          type: integer
          format: int64
        orientation:
          $ref: '#/components/schemas/Photos.PhotoOrientation'
        software:
          type: string
        firmware:
          type: string
    Photos.File:
      type: object
      required:
        - fileId
        - photoId
        - fileType
        - fileHash
        - fileName
      properties:
        fileId:
          type: string
        photoId:
          type: string
        fileType:
          type: string
        fileHash:
          type: string
        fileName:
          type: string
    Photos.Photo:
      type: object
      required:
        - photoId
        - name
        - importedAt
        - dateTimeOriginal
        - previewUrl
        - thumbnailUrl
        - exifData
        - fileTypes
        - files
      properties:
        photoId:
          type: string
        name:
          type: string
        importedAt:
          type: string
        dateTimeOriginal:
          type: string
          format: date-time
        previewUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        exifData:
          $ref: '#/components/schemas/Photos.ExifData'
        fileTypes:
          type: array
          items:
            type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/Photos.File'
    Photos.PhotoItem:
      type: object
      required:
        - photoId
        - name
        - importedAt
        - dateTimeOriginal
        - previewUrl
        - thumbnailUrl
      properties:
        photoId:
          type: string
        name:
          type: string
        importedAt:
          type: string
        dateTimeOriginal:
          type: string
          format: date-time
        previewUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
    Photos.PhotoListResponse:
      type: object
      required:
        - items
        - total
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Photos.PhotoItem'
        total:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
    Photos.PhotoOrientation:
      type: number
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
      x-go-type: int64
      x-enum-varnames:
        - ExifOrientationNone
        - ExifOrientationHorizontal
        - ExifOrientationRotate180
        - ExifOrientationVertical
        - ExifOrientationHorizontalRotate270
        - ExifOrientationRotate90
        - ExifOrientationHorizontalRotate90
        - ExifOrientationRotate270
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session
servers:
  - url: http://localhost:8080
    description: local
    variables: {}